# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/wells": {
      "get": {
        "description": "Paginated well catalog. Supports filters by operator/formation/basin/province/concession and a substring search on `sigla`.",
        "operationId": "WellsController_list_v1",
        "parameters": [
          {
            "name": "operator",
            "required": false,
            "in": "query",
            "description": "Filter by operator slug.",
            "schema": {
              "example": "ypf",
              "type": "string"
            }
          },
          {
            "name": "formation",
            "required": false,
            "in": "query",
            "description": "Filter by formation slug.",
            "schema": {
              "example": "vaca_muerta",
              "type": "string"
            }
          },
          {
            "name": "basin",
            "required": false,
            "in": "query",
            "description": "Filter by basin (matches CSV exactly).",
            "schema": {
              "example": "NEUQUINA",
              "type": "string"
            }
          },
          {
            "name": "province",
            "required": false,
            "in": "query",
            "description": "Filter by province.",
            "schema": {
              "example": "Neuquén",
              "type": "string"
            }
          },
          {
            "name": "concession",
            "required": false,
            "in": "query",
            "description": "Filter by concession (matches CSV exactly).",
            "schema": {
              "example": "BANDURRIA SUR",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring search on `sigla` (well name).",
            "schema": {
              "example": "BAN",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 50,
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/WellDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationMetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta",
                    "pagination"
                  ]
                }
              }
            }
          }
        },
        "summary": "List wells",
        "tags": [
          "wells"
        ]
      }
    },
    "/api/v1/wells/{id}": {
      "get": {
        "description": "Well metadata plus a `latest_production` block with the most recent monthly fact for this well.",
        "operationId": "WellsController_detail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Internal `well_id` (numeric string).",
            "schema": {
              "example": "93863",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/WellDetailDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Well ID not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Well detail",
        "tags": [
          "wells"
        ]
      }
    },
...
    "/api/v1/geo/wells": {
      "get": {
        "description": "Returns **raw GeoJSON** (no `data` / `meta` envelope) so the response can be fed directly into Mapbox / MapLibre / Leaflet.\n\nUse `bbox=west,south,east,north` for viewport-based loading.\n\nWells without coordinates are excluded. The `limit` caps total features per request (max 1000).",
        "operationId": "GeoController_wells_v1",
        "parameters": [
          {
            "name": "operator",
            "required": false,
            "in": "query",
            "schema": {
              "example": "ypf",
              "type": "string"
            }
          },
          {
            "name": "formation",
            "required": false,
            "in": "query",
            "schema": {
              "example": "vaca_muerta",
              "type": "string"
            }
          },
          {
            "name": "basin",
            "required": false,
            "in": "query",
            "schema": {
              "example": "NEUQUINA",
              "type": "string"
            }
          },
          {
            "name": "province",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Neuquén",
              "type": "string"
            }
          },
          {
            "name": "bbox",
            "required": false,
            "in": "query",
            "description": "Bounding box `west,south,east,north` in WGS84 decimal degrees. Use for viewport-based loading.",
            "schema": {
              "example": "-70,-39,-68,-37",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "default": 1000,
              "example": 1000,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "GeoJSON FeatureCollection of well points.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoWellFeatureCollectionDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid bbox format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Wells as a GeoJSON FeatureCollection",
        "tags": [
          "geo"
        ]
      }
    },
    "/api/v1/geo/wells/{id}": {
      "get": {
        "description": "Returns raw GeoJSON. 404 if the well has no recorded coordinates.",
        "operationId": "GeoController_well_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "10000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "GeoJSON Feature for one well.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoWellFeatureDto"
                }
              }
            }
          },
          "404": {
            "description": "Well missing or has no geometry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Single well as a GeoJSON Feature",
        "tags": [
          "geo"
        ]
      }
    },
...
  },
  "info": {
    "title": "Petroldata.ar API",
    "description": "REST API over Argentina Secretaría de Energía oil & gas production data (focus: Vaca Muerta).\n\n## Response envelope\n\nEvery non-GeoJSON endpoint returns:\n\n```json\n{\n  \"data\": ...,\n  \"meta\": {\n    \"source\": \"Secretaría de Energía / datos.energia.gob.ar\",\n    \"dataset\": \"Producción de petróleo y gas por pozo\",\n    \"license\": \"CC-BY-4.0\",\n    \"last_source_update\": \"2026-04-01\",\n    \"last_ingested_at\": \"2026-05-27T14:30:00.000Z\",\n    \"vaca_muerta_filter\": \"formation + unconventional + sub-tipo\"\n  },\n  \"pagination\": { \"page\": 1, \"limit\": 50, \"total\": 1000 }\n}\n```\n\n`pagination` is only present on paginated list endpoints. GeoJSON endpoints (`/geo/*`) return raw GeoJSON.\n\n## Errors\n\n```json\n{ \"error\": { \"code\": \"NOT_FOUND\", \"message\": \"Well not found: 99999\" } }\n```\n\n## Units\n\n- `oil_m3` — cubic metres of oil\n- `oil_bbl` — barrels of oil (1 m³ ≈ 6.2898 bbl)\n- `oil_bbl_d` — barrels per day (oil_bbl ÷ days in month)\n- `gas_thousand_m3` — thousands of cubic metres of gas (Mm³)\n- `gas_mcf` — thousand cubic feet of gas\n- `gas_mmcf_d` — million cubic feet per day\n- `gas_mm3_d` — million m³ per day\n- `boe` — barrels of oil equivalent (oil_bbl + gas_boe)\n\n## Vaca Muerta flag\n\n`vm_combined = formation_vaca_muerta AND unconventional AND sub_resource_type ∈ {Shale, Shale Oil, …}` — pre-computed in the Python pipeline.",
...
  "servers": [
...
  "components": {
    "schemas": {
...
      "MetaDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "example": "Secretaría de Energía / datos.energia.gob.ar"
          },
          "dataset": {
            "type": "string",
            "example": "Producción de petróleo y gas por pozo"
          },
          "license": {
            "type": "string",
            "example": "CC-BY-4.0"
          },
          "last_source_update": {
            "type": "object",
            "example": "2026-04-01",
            "nullable": true
          },
          "last_ingested_at": {
            "type": "string",
            "example": "2026-05-27T14:30:00.000Z"
          },
          "vaca_muerta_filter": {
            "type": "string",
            "example": "formation + unconventional + sub-tipo"
          }
        },
        "required": [
          "source",
          "dataset",
          "license",
          "last_source_update",
          "last_ingested_at",
          "vaca_muerta_filter"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 50
          },
          "total": {
            "type": "number",
            "example": 1000
          }
        },
        "required": [
          "page",
          "limit",
          "total"
        ]
      },
...
      "ApiErrorPayloadDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "Well not found: 99999"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorPayloadDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "WellDto": {
        "type": "object",
        "properties": {
          "well_id": {
            "type": "string",
            "example": "93863"
          },
          "sigla": {
            "type": "string",
            "example": "ACB.a-402"
          },
          "operator_slug": {
            "type": "string",
            "example": "ypf"
          },
          "operator_name": {
            "type": "string",
            "example": "YPF S.A."
          },
          "formation_slug": {
            "type": "string",
            "example": "grupo_chubut"
          },
          "province": {
            "type": "string",
            "example": "Chubut"
          },
          "basin": {
            "type": "string",
            "example": "GOLFO SAN JORGE"
          },
          "concession": {
            "type": "string",
            "example": "MANANTIALES BEHR"
          },
          "yacimiento": {
            "type": "string",
            "example": "MANANTIALES BEHR"
          },
          "well_type": {
            "type": "string",
            "example": "Petrolífero"
          },
          "extraction_type": {
            "type": "string",
            "example": "Bombeo Mecánico"
          },
          "status_code": {
            "type": "string",
            "example": "En Producción Efectiva"
          },
          "resource_type": {
            "type": "string",
            "example": "CONVENCIONAL"
          },
          "sub_resource_type": {
            "type": "string",
            "example": "No informado"
          },
          "depth_m": {
            "type": "object",
            "example": 2300,
            "nullable": true
          },
          "latitude": {
            "type": "object",
            "example": -45.62363321,
            "nullable": true
          },
          "longitude": {
            "type": "object",
            "example": -67.8618809,
            "nullable": true
          }
        },
        "required": [
          "well_id",
          "sigla",
          "operator_slug",
          "operator_name",
          "formation_slug",
          "province",
          "basin",
          "concession",
          "yacimiento",
          "well_type",
          "extraction_type",
          "status_code",
          "resource_type",
          "sub_resource_type",
          "depth_m",
          "latitude",
          "longitude"
        ]
      },
      "WellLatestProductionDto": {
        "type": "object",
        "properties": {
          "date_month": {
            "type": "string",
            "example": "2026-04-01"
          },
          "oil_m3": {
            "type": "number",
            "example": 152.45
          },
          "oil_bbl_d": {
            "type": "number",
            "example": 31.95
          },
          "gas_thousand_m3": {
            "type": "number",
            "example": 84.32
          },
          "gas_mmcf_d": {
            "type": "number",
            "example": 0.099
          },
          "boe": {
            "type": "number",
            "example": 1212.7
          },
          "vm_combined": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "date_month",
          "oil_m3",
          "oil_bbl_d",
          "gas_thousand_m3",
          "gas_mmcf_d",
          "boe",
          "vm_combined"
        ]
      },
      "WellDetailDto": {
        "type": "object",
        "properties": {
          "well_id": {
            "type": "string",
            "example": "93863"
          },
          "sigla": {
            "type": "string",
            "example": "ACB.a-402"
          },
          "operator_slug": {
            "type": "string",
            "example": "ypf"
          },
          "operator_name": {
            "type": "string",
            "example": "YPF S.A."
          },
          "formation_slug": {
            "type": "string",
            "example": "grupo_chubut"
          },
          "province": {
            "type": "string",
            "example": "Chubut"
          },
          "basin": {
            "type": "string",
            "example": "GOLFO SAN JORGE"
          },
          "concession": {
            "type": "string",
            "example": "MANANTIALES BEHR"
          },
          "yacimiento": {
            "type": "string",
            "example": "MANANTIALES BEHR"
          },
          "well_type": {
            "type": "string",
            "example": "Petrolífero"
          },
          "extraction_type": {
            "type": "string",
            "example": "Bombeo Mecánico"
          },
          "status_code": {
            "type": "string",
            "example": "En Producción Efectiva"
          },
          "resource_type": {
            "type": "string",
            "example": "CONVENCIONAL"
          },
          "sub_resource_type": {
            "type": "string",
            "example": "No informado"
          },
          "depth_m": {
            "type": "object",
            "example": 2300,
            "nullable": true
          },
          "latitude": {
            "type": "object",
            "example": -45.62363321,
            "nullable": true
          },
          "longitude": {
            "type": "object",
            "example": -67.8618809,
            "nullable": true
          },
          "latest_production": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/WellLatestProductionDto"
              }
            ]
          }
        },
        "required": [
          "well_id",
          "sigla",
          "operator_slug",
          "operator_name",
          "formation_slug",
          "province",
          "basin",
          "concession",
          "yacimiento",
          "well_type",
          "extraction_type",
          "status_code",
          "resource_type",
          "sub_resource_type",
          "depth_m",
          "latitude",
          "longitude",
          "latest_production"
        ]
      },
...
      "GeoPointGeometryDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "Point"
          },
          "coordinates": {
            "example": [
              -69.36352,
              -37.40809
            ],
            "description": "[longitude, latitude] in WGS84, per GeoJSON spec.",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "type",
          "coordinates"
        ]
      },
      "GeoWellPropertiesDto": {
        "type": "object",
        "properties": {
          "well_id": {
            "type": "string",
            "example": "10000"
          },
          "sigla": {
            "type": "string",
            "example": "SJ.Nq.ET\"E\".x-1"
          },
          "operator_slug": {
            "type": "string",
            "example": "chevron"
          },
          "operator_name": {
            "type": "string",
            "example": "CHEVRON ARGENTINA S.R.L."
          },
          "formation_slug": {
            "type": "string",
            "example": "huitr_n"
          },
          "basin": {
            "type": "string",
            "example": "NEUQUINA"
          },
          "province": {
            "type": "string",
            "example": "Neuquén"
          },
          "concession": {
            "type": "string",
            "example": "EL TRAPIAL ESTE"
          },
          "yacimiento": {
            "type": "string",
            "example": "EL TRAPIAL"
          },
          "well_type": {
            "type": "string",
            "example": "Petrolífero"
          },
          "status_code": {
            "type": "string",
            "example": "En Estudio"
          },
          "depth_m": {
            "type": "object",
            "example": 1315,
            "nullable": true
          }
        },
        "required": [
          "well_id",
          "sigla",
          "operator_slug",
          "operator_name",
          "formation_slug",
          "basin",
          "province",
          "concession",
          "yacimiento",
          "well_type",
          "status_code",
          "depth_m"
        ]
      },
      "GeoWellFeatureDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "Feature"
          },
          "geometry": {
            "$ref": "#/components/schemas/GeoPointGeometryDto"
          },
          "properties": {
            "$ref": "#/components/schemas/GeoWellPropertiesDto"
          }
        },
        "required": [
          "type",
          "geometry",
          "properties"
        ]
      },
      "GeoWellFeatureCollectionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "FeatureCollection"
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeoWellFeatureDto"
            }
          }
        },
        "required": [
          "type",
          "features"
        ]
      },
...
    }
  }
}